Admin logic is done
Admin.Razo: Added som style to the buttons and card
AdminAddEven: The UI and form is done to create a new event.
ConfirmationDialog component: Deleted the header.
EventService: Added an exception handler.
EventModel: Changed the Tags from null to a new list of tags to be able to use it in AdminAddEvent page.